ページメニュー：英語に切り替えて、ページ追加すると日本語に戻るバグを修正。英語に切り替えて、コミュニティー編集、追加も同様。

+++ lib/Nc/Plugin/Page/Controller/PageController.php
@@ -70,6 +70,9 @@ class PageController extends PageAppController {
 
 		$activeLang = isset($this->request->named['active_lang']) ? $this->request->named['active_lang'] : null;
 		$sessLang = $this->Session->read(NC_CONFIG_KEY.'.language');
+		if(!isset($activeLang)) {
+			$activeLang = $this->Session->read(NC_SYSTEM_KEY.'.page_menu.activeLang');
+		}
 
 		$this->Session->write(NC_SYSTEM_KEY.'.page_menu.lang', $sessLang);
 		parent::beforeFilter();
@@ -77,6 +80,7 @@ class PageController extends PageAppController {
 		if(isset($activeLang) && isset($languages[$activeLang])) {
 			Configure::write(NC_CONFIG_KEY.'.'.'language', $activeLang);
 			$this->Session->write(NC_CONFIG_KEY.'.language', $activeLang);
+			$this->Session->write(NC_SYSTEM_KEY.'.page_menu.activeLang', $activeLang);
 		}
 
 		if ($this->action == 'index') {
@@ -119,8 +123,9 @@ class PageController extends PageAppController {
 	public function afterFilter()
 	{
 		parent::afterFilter();
-		if($this->action != 'index') {
+		if($this->action != 'index' && $this->action != 'close') {
 			$this->Session->delete(NC_SYSTEM_KEY.'.page_menu.lang');
+			$this->Session->delete(NC_SYSTEM_KEY.'.page_menu.activeLang');
 		}
 		$preLang = $this->Session->read(NC_SYSTEM_KEY.'.page_menu.lang');
 		if(isset($preLang)) {
